phase_2_full_stack_web_application/backend/requirements.txt
alembic>=1.17.2
fastapi>=0.124.4
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.11
pydantic-settings>=2.12.0
pydantic[email]>=2.12.5
python-jose[cryptography]>=3.5.0
python-multipart>=0.0.20
sqlmodel>=0.0.27
uvicorn>=0.38.0
httpx>=0.28.1
pytest>=9.0.2
pytest-asyncio>=1.3.0
python-dotenv>=1.0.1
bcrypt>=3.1.3,<4.0.0
cryptography>=45.0.0